% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comoDistList}
\alias{comoDistList}
\title{Comonotonize a list of distributions}
\usage{
comoDistList(lisOfDists, regridMethod = "lr", N = 256L)
}
\arguments{
\item{lisOfDists}{A list of PMFs. 

  Each PMF is a 2-column data frame \code{(support, probability)}. 
  \code{support} should be regular (equal step sizes). \code{probability} 
  should be nonnegative. 

  A degenerate PMF must be a 1-row data frame. A degenerate PMF represented as 
  \code{data.frame(val = c(0, 1, 2), P = c(0, 0, 1))} instead of 
  \code{data.frame(val = 2, P = 1)} can lead to runtime error.}

\item{regridMethod}{Regriding method:

  \code{"lr"}: linear regrid. Default.

  \code{"r4"}: four-point regrid.}

\item{N}{Maximum points allowed on the result PMF's support. Default 256.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Comonotonize a list of PMFs sequentially.
}
\details{
See \code{\link{como}}.
}
\examples{
# Generate a list of 100 PMFs.
distlist = lapply(1:100, function(i)
{
  supportSize = sample(200, 1)
  supportRange = sort(runif(2))
  support = seq(supportRange[1], supportRange[2], len = supportSize)
  X = data.frame(val = support, P = runif(supportSize))
  X$P = X$P / sum(X$P); X
})


comoAggDist = keyALGs::comoDistList(distlist, regridMethod = "r4")
allSDs = unlist(lapply(distlist, function(x) sqrt(keyALGs::Var(x))))
aggVarIfAllPerfectlyCorrelated = sum(allSDs) ^ 2
comoAggDistVar = keyALGs::Var(comoAggDist)


theoreticalMean = sum(unlist(lapply(distlist, function(x) keyALGs::Mean(x))))
cat("Theoretical aggregate mean = ", theoreticalMean, "\n",
    "Comonotonic aggregate mean = ", keyALGs::Mean(comoAggDist), "\n",
    "Variance upper bound of the aggregate PMF (if all random variables are
perfectly correlated) = ", aggVarIfAllPerfectlyCorrelated, "\n",
    "Comonotonic aggregate variance = ", comoAggDistVar, "\n",
    "Comonotonic aggregate variance may exceed the theoretical upper bound due
to variance inflation from regriding.", 
    sep = "")
}
