\name{comoJoint2D}
\alias{comoJoint2D}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Comonotonic joint
}
\description{
Compute joint distribution of comonotonization.
}
\usage{
comoJoint2D(X, Y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
A 2-column data frame as the PMF.
}
  \item{Y}{
A 2-column data frame as the PMF.
}
}
\details{

}
\value{
A 3-column data frame (x, y, prob).
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
