% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convBt}
\alias{convBt}
\title{Brute-force convolution}
\usage{
convBt(X, Y, eps = 1e-10)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}

\item{Y}{Same as \code{X}.}

\item{eps}{Numeric threshold to define equality between double-precision 
floats: if \code{|a - b| < eps}, then \code{a} and \code{b} are treated as 
equal. Default \code{1e-10}.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Brute-force convolve two PMFs where each PMF's support can be fully irregular.
}
\examples{
X = data.frame(val = sort(runif(100)), P = runif(100))
X$P = X$P / sum(X$P)
Y = data.frame(val = sort(runif(200)), P = runif(200))
Y$P = Y$P / sum(Y$P)
XplusY = keyALGs::convBt(X, Y, eps = 1e-10)


cat("Result PMF has only nonnegative probabilities = ",
    min(XplusY$P) >= 0, "\n",
    "Sum of result PMF's probabilities = ",
    sum(XplusY$P), "\n",
    "Theoretical mean of the result PMF = ",
    keyALGs::Mean(X) + keyALGs::Mean(Y), "\n",
    "Actual mean of the result PMF = ",
    keyALGs::Mean(XplusY), "\n",
    "Theoretical variance of the result PMF = ",
    keyALGs::Var(X) + keyALGs::Var(Y), "\n",
    "Actual variance of the result PMF = ",
    keyALGs::Var(XplusY), "\n", sep = "")
}
