% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{emp}
\alias{emp}
\title{Compute empirical distribution.}
\usage{
emp(x, N = 64L, rgMethod = "lr")
}
\arguments{
\item{x}{A numeric vector of samples.}

\item{N}{An integer. Support size of the result PMF. Default 64.}

\item{rgMethod}{Regriding method:

\code{"lr"}: linear regrid. Default.

\code{"r4"}: four-point regrid.

\code{"r3"}: three-point regrid. For author's personal use.

\code{"r4g"}: four-point regrid with Gaussian kernel. For author's personal use.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Compute the empirical PMF of a sample set.
}
\details{
The function creates a PMF with \code{x} as the support and
\code{1/length(x)} as the probabilities, and then regrids the PMF onto a 
regular grid of \code{N} points within \code{[min(x), max(x)]}.
}
\examples{
# x = sort(rnorm(100))
pX = keyALGs::emp(x, N = 30, rgMethod = "r4")
plot(pX, type = "h")
}
