\name{LHS}
\alias{LHS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Latin hypercube sampling
}
\description{
Latin hypercube sampling from a discrete PDF.
}
\usage{
LHS(pdf, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pdf}{
PDF as a 2-column data frame. The first column is values and the second column is probabilities.
}
  \item{N}{
Sample size.
}
}
\details{
}
\value{
A numeric vector as the sample.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
