% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.r
\name{LMMdiscreteToDiscrete}
\alias{LMMdiscreteToDiscrete}
\title{Regriding via local (up to 2nd) moment matching}
\usage{
LMMdiscreteToDiscrete(X, ngd)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}

\item{ngd}{A sorted numeric vector as the new grid: \code{ngd[1] <= support[1]}
AND \code{ngd[length(ngd)] >= support[length(support)]}.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Regrid a PMF via local moment matching (LMM). The algorithm usually but not 
always preserves the PMF's second moment.
}
\details{
See \href{https://www.mdpi.com/2227-9091/7/2/54/htm}{Direct and 
Hierarchical Models for Aggregating Spatially Dependent Catastrophe Risks}.
\code{\link{LMMdiscreteToDiscrete}} would more likely inflate variance than 
four-point regriding \code{\link{rglr4split}}.
}
\examples{
X = data.frame(val = seq(0, 16, len = 1000), P = dgamma(
  seq(0, 16, len = 1000), shape = 2, scale = 2)) # Input PMF's column 
# names do not have to be "val" and "P".


X$P = X$P / sum(X$P)
newGrid = seq(0, 16, len = 64)
Xnew = keyALGs::LMMdiscreteToDiscrete(X, newGrid)


# Means:
keyALGs::Mean(X); keyALGs::Mean(Xnew)
# 3.957082
# 3.957082


# Variances:
keyALGs::Var(X); keyALGs::Var(Xnew)
# 7.396614
# 7.396614
}
