% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mean}
\alias{Mean}
\title{Distribution mean}
\usage{
Mean(X)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}
}
\value{
Mean of the PMF
}
\description{
Compute the mean of a PMF.
}
