% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NhistBins.R
\name{NhistBins}
\alias{NhistBins}
\title{Optimal number of bins in a histogram}
\usage{
NhistBins(x)
}
\arguments{
\item{x}{A vector as the sample.}
}
\value{
An integer.
}
\description{
Compute the optimal number of bins in a histogram for
summarizing a sample based on the Freedman-Diaconis rule.
}
