% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printExcel.R
\name{printExcel}
\alias{printExcel}
\title{Print Excel sheet.}
\usage{
printExcel(
  excelBookFullPath,
  sheetNames,
  printArea = "A1:B2",
  pdfFullPath = "sample.pdf"
)
}
\arguments{
\item{excelBookFullPath}{Full path to the Excel book. Relative path will not
work since Excel does not recognize your current Python working directory.}

\item{printArea}{A character string in the format of "A1:B2" to identify the
top-left and bottom-right cells.}

\item{pdfFullPath}{Full path to the printed pdf.}

\item{sheetName}{A character string. Name of the sheet in the Excel book.}
}
\value{
Nothing.
}
\description{
Print a rectangle block of cells in a sheet in an Excel book.
}
\details{
The function uses \code{\link[reticulate]{source_python}} to 
source "excelPrint.py" in /inst/python.
}
