% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rglr}
\alias{rglr}
\title{Linear regriding}
\usage{
rglr(X, ngd)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}

\item{ngd}{A sorted numeric vector as the new grid: \code{ngd[1] <= support[1]}
AND \code{ngd[length(ngd)] >= support[length(support)]}.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Linear regrid a PMF
}
\details{
See \href{https://www.mdpi.com/2227-9091/7/2/54/htm}{Direct and
Hierarchical Models for Aggregating Spatially Dependent Catastrophe Risks}.
}
\examples{
X = data.frame(val = sort(runif(100)), P = runif(100)) # Input PMF's column
# names do not have to be "val" and "P".
X$P = X$P / sum(X$P)
newSupport = sort(c(runif(28), X$val[1], X$val[nrow(X)]))
Y = keyALGs::rglr(X, newSupport)


# Plot the two PMFs.
ylim = c(0, max(max(X$P), max(Y$P)))
par(mfrow = c(1, 2))
plot(X, type = "h", bty = "L", ylim = ylim)
legend("top", legend = expression(p[X]), bty = "n")
plot(Y, type = "h", bty = "L", ylim = ylim, ylab = "", xlab = "")
legend("top", legend = expression(p[Y]), bty = "n")
}
