% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rglr4split}
\alias{rglr4split}
\title{Four-point regriding}
\usage{
rglr4split(X, ngd)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}

\item{ngd}{A sorted numeric vector as the new grid: \code{ngd[1] <= support[1]}
AND \code{ngd[length(ngd)] >= support[length(support)]}.}
}
\value{
A 2-column data frame as the regrided PMF. Column names are
\code{"val"} and \code{"P"}.
}
\description{
Regrid a PMF using four-point regriding. The algorithm usually but not always 
preserves the PMF's second moment.
}
\details{
See \href{https://www.mdpi.com/2227-9091/7/2/54/htm}{Direct and
Hierarchical Models for Aggregating Spatially Dependent Catastrophe Risks}.
}
\examples{
X = data.frame(val = sort(runif(100)), P = runif(100))  # Input PMF's column 
# names do not have to be "val" and "P".
X$P = X$P / sum(X$P)
newSupport = sort(c(runif(28), X$val[1], X$val[nrow(X)]))
linearRegridedX = keyALGs::rglr(X, newSupport)
fourPointRegridedX = keyALGs::rglr4split(X, newSupport)


# Linear regrid inflated variance. Four-point regrid preserved variance. 
cat("Mean(X), Var(X) = ", 
    keyALGs::Mean(X), ", ", keyALGs::Var(X), "\n",
    "Mean(linearRegridedX), Var(linearRegridedX) = ", 
    keyALGs::Mean(linearRegridedX), ", ", keyALGs::Var(linearRegridedX), "\n",
    "Mean(fourPointRegridedX), Var(fourPointRegridedX) = ", 
    keyALGs::Mean(fourPointRegridedX), ", ", keyALGs::Var(fourPointRegridedX), 
    sep = "")
}
