% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rglr4splitDetails}
\alias{rglr4splitDetails}
\title{Four-point regriding details}
\usage{
rglr4splitDetails(X, ngd)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}

\item{ngd}{A sorted numeric vector as the new grid: \code{ngd[1] <= support[1]}
AND \code{ngd[length(ngd)] >= support[length(support)]}.}
}
\value{
A list of 3:
\itemize{
  \item{\code{$RTL}}{ TRUE if the algorithm retreated to linear regriding.}
  \item{\code{$Nredist}}{ An integer. The number of redistributions of negative 
  probabilities in Stage II.}
  \item{\code{$resultDistr}}{ A 2-column data frame 
  \code{(support, probability)} as the regrided PMF.}
 }
}
\description{
Regrid a PMF using four-point regriding. The algorithm usually but not always 
preserves the PMF's second moment. In addition to the regrided PMF, the function 
returns some process information.
}
\details{
See \href{https://www.mdpi.com/2227-9091/7/2/54/htm}{Direct and
Hierarchical Models for Aggregating Spatially Dependent Catastrophe Risks}.
}
\examples{
set.seed(42)
X = data.frame(val = sort(runif(30)), P = c(runif(14), 300, runif(15))) # Input 
# PMF's column names do not have to be "val" and "P".


X$P = X$P / sum(X$P)
plot(X, type = "h") # Made a close-to-degenerate PMF to increase the chance 
# of algorithm entering Stage II. See the publication referenced in Details.


newSupport = sort(c(X$val[1] - runif(1) * 0.05,
                    runif(20), X$val[nrow(X)] + runif(1) * 0.05))


fourPointRegridDetails = keyALGs::rglr4splitDetails(X, newSupport)


fourPointRegridDetails




}
