% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Var}
\alias{Var}
\title{Distribution variance}
\usage{
Var(X)
}
\arguments{
\item{X}{A 2-column data frame \code{(support, probability)} as the PMF.
Elements in \code{support} must be sorted and have no duplicates.
\code{probability} should be nonnegative.}
}
\value{
Variance of the PMF
}
\description{
Compute the variance of a PMF.
}
